+++ x-pack/plugins/apm/public/components/app/ErrorGroupDetails/DetailView/__test__/DetailView.test.js
@@ -5,7 +5,7 @@
  */
 
 import React from 'react';
-import DetailView from '../index';
+import { DetailView } from '../index';
 import props from './props.json';
 import { shallow } from 'enzyme';
